To be able to resend, stop all voice actions without deleting.

+++ vector/src/main/java/im/vector/app/features/home/room/detail/composer/VoiceMessageHelper.kt
@@ -217,10 +217,12 @@ class VoiceMessageHelper @Inject constructor(
         playbackTicker = null
     }
 
-    fun stopAllVoiceActions() {
+    fun stopAllVoiceActions(deleteRecord: Boolean = true) {
         stopRecording()
         stopPlayback()
-        deleteRecording()
+        if (deleteRecord) {
+            deleteRecording()
+        }
         playbackTracker.clear()
     }
 }
